refactor utility-functions
