+++ src/lib/utils/capitalize.ts
@@ -1,0 +1,5 @@
+/* capitalize given string using RegEx */
+const capitalize = (string: string) =>
+  string?.replace(/^\w/, (ch) => ch.toUpperCase());
+
+export default capitalize;
